[rename] selectDayList.js에서 selectDayHome.js으로 파일명 변경 - makeSelectDayItems와 관련된 함수를 새로운 모듈로 분리했다.
