+++ public/JS/components/selectDay/selectDayHome.js
@@ -1,3 +1,4 @@
+import makeSelectDayItems from './selectDayItem.js'
 import {makeWebtoonList} from '../webtoonList.js'
 import {getData} from '../../utility.js'
 
@@ -24,19 +25,4 @@ function makeSelectDayHome(num, data) {
   `;
 }
 
-function makeSelectDayItems() {
-  const day = ['월', '화', '수', '목', '금', '토', '일', '완결'];
-  const dataSet = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun', 'end'];
-
-  return day.reduce((acc, cur, index) => acc + makeSelectDayItem(cur, dataSet, index), '');
-}
-
-function makeSelectDayItem(day, dataSet, index) {
-  return `
-    <li class="day" data-day="${dataSet[index]}">
-      <button type="button" class="day__btn">${day}</button>
-    </li>
-  `
-}
-
-export {renderSelectDayHome, makeSelectDayItems}
+export {renderSelectDayHome}
